python/resinCounter/requirements.txt
browser-cookie3==0.19.1
certifi==2023.5.7
charset-normalizer==3.2.0
genshinstats==1.4.11.3
idna==3.4
lz4==4.3.2
pycryptodomex==3.18.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
